back-end/pyproject.toml
[tool.poetry]
name = "trading-wheels-server"
version = "0.1.0"
description = ""
authors = ["et-codes <104942211+et-codes@users.noreply.github.com>"]
license = "LICENSE"
readme = "README.md"
packages = [{include = "trading_wheels_server"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.2"
flask-cors = "^3.0.10"
requests = "^2.28.1"
python-dotenv = "^0.21.0"
pyjwt = "^2.6.0"
gunicorn = "^20.1.0"
flask-sqlalchemy = "^3.0.2"
psycopg2 = "^2.9.5"
flask-migrate = "^3.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
